pyproject.toml
[project]
name = "playwright-percy-snapshot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "hemanth-kumar-j", email = "hemanthkumar.jhemanth@rediffmail.com" }
]
dependencies = [
    "playwright>=1.51.0",
    "pytest>=8.3.5",
    "pytest-html>=4.1.1",
    "black>=25.1.0",
    "python-dotenv>=1.1.1",
    "percy-playwright>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/playwright_percy_snapshot"]

[tool.rye.scripts]
playwright-test = "python browser_test.py --browser {args}"
